[MM-204] Review fixes: 1. Renamed constants. 2. Created a util function for getting truncated text for title and description of tooltip.
