--- webapp/src/utils/regex_utils.ts
@@ -1,4 +1,0 @@
-export function validateGitlabURL(url: string): boolean {
-    const gitlabRegexPattern = /https?:\/\/(www\.)?.*\/([\w.?-]+)\/([\w-]+)\/-\/([\w-]+)\/([\d-]+$)/g;
-    return gitlabRegexPattern.test(url);
-}
