--- webapp/src/utils/url_utils.ts
@@ -1,7 +1,0 @@
-export const isValidUrl = (urlString: string): boolean => {
-    try {
-        return Boolean(new URL(urlString));
-    } catch (e) {
-        return false;
-    }
-};
